# old "maddylogin" packages fixed to "rebase jumpers", also new icon added for app and toolbar from "MaddyLogin" to "Where's My Stuff?"
