+++ app/src/main/java/com/example/RB/rebasejumpers/Logout_Screen.java
@@ -1,0 +1,31 @@
+package com.example.RB.rebasejumpers;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
+
+public class Logout_Screen extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_logout__screen);
+
+        Button logoutButton = (Button) findViewById(R.id.logout_button);
+        logoutButton.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View view) {
+                startActivity(new Intent(Logout_Screen.this, LoginActivity.class));
+            }
+        });
+
+        Button itemListButton = (Button) findViewById(R.id.item_list_button);
+        itemListButton.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View view) {
+                startActivity(new Intent(Logout_Screen.this, ItemView.class));
+            }
+        });
+    }
+
+}
