--- app/src/main/java/com/example/maddy/maddylogin/Logout_Screen.java
@@ -1,34 +1,0 @@
-package com.example.maddy.maddylogin;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.widget.Button;
-
-public class Logout_Screen extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_logout__screen);
-
-        Button logoutButton = (Button) findViewById(R.id.logout_button);
-        logoutButton.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View view) {
-                startActivity(new Intent(Logout_Screen.this, LoginActivity.class));
-            }
-        });
-
-        Button itemListButton = (Button) findViewById(R.id.item_list_button);
-        itemListButton.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View view) {
-                startActivity(new Intent(Logout_Screen.this, ItemView.class));
-            }
-        });
-    }
-
-}
