+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.maddy.maddylogin.LoginActivity"
+    tools:context="com.example.RB.rebasejumpers.LoginActivity"
     android:orientation="vertical">
 
     <!-- Login progress -->
