+++ app/src/main/res/layout/content_registration1.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.maddy.maddylogin.RegistrationActivity1"
+    tools:context="com.example.RB.rebasejumpers.RegistrationActivity1"
     tools:showIn="@layout/activity_registration1">
 
 </android.support.constraint.ConstraintLayout>
