+++ local.properties
@@ -7,5 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Fri Jun 23 15:29:26 EDT 2017
-sdk.dir=/home/will/Android/Sdk
+#Sat Jun 24 11:46:38 EDT 2017
+sdk.dir=/Users/maddy/Library/Android/sdk
